+++ CHANGELOG.rst
@@ -19,12 +19,14 @@ Added
 - Added `VectorSignal1D` class to handle 1 dimensional signals
 - Added kwargs to find_beam_offset_cross_correlation allowing for parameters
 to be passed to `phase_cross_correlation` (#907)
+- Added `LazyVirtualDarkField` signal
 
 Changed
 -------
 - Sklearn is now required to be on the 1.0 series.
 - Changed `set_signal_dimension` to `Hyperspy.signals.BaseSignal.transpose`
 - Moved code from `VectorSignal` to `VectorSignal2D`.  Change is more inline with stated dimensions
+- `VectorSignal` pixel_calibration deprecated and replaced with scales.
 - Fixed bugs resulting from API change in hyperspy/hyperspy#3045. Markers explicitly initialized
 - DiffractionVectors.get_diffraction_pixels_map returns a ragged signal
 - VirtualDarkFieldImage.get_vdf_segment changed to properly handle setting of axes
@@ -38,6 +40,8 @@ Fixed
 - Bugfix related to Numpy 1.24.0. Strict array creation with dtype=object is needed
   for ragged arrays. (#880 & #881)
 - Bug fix for doubling of inplane rotation in template matching.  (#905 & #853)
+- Bug fix for filtering vectors using a basis and DBSCAN
+- Bug fix for passing vector attributes when signal is copied or changed.
 
 
 2022-15-06 - version 0.14.2